Add GIDD page for config
