+++ src/components/tables/ReportsTable/styles.css
@@ -1,0 +1,10 @@
+.content {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+
+    .table {
+        width: 100%;
+        overflow-x: auto;
+    }
+}
