+++ src/views/Gidd/GiddSettings/styles.css
@@ -1,0 +1,13 @@
+.form {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    gap: var(--spacing-medium);
+
+    .form-buttons{
+        display: flex;
+        flex-direction: row;
+        justify-content: flex-end;
+        gap: var(--spacing-medium);
+    }
+}
